+++ data/StoreRecord.ts
@@ -6,7 +6,7 @@
  */
 import {PlainObject} from '@xh/hoist/core';
 import {throwIf} from '@xh/hoist/utils/js';
-import {isNil, flatMap, isMatch, isEmpty} from 'lodash';
+import {isNil, flatMap, isMatch, isEmpty, pickBy} from 'lodash';
 import {Store} from './Store';
 import {ValidationState} from './validation/ValidationState';
 import {RecordValidator} from './impl/RecordValidator';
@@ -73,7 +73,7 @@ export class StoreRecord {
 
     /** True if the StoreRecord has been modified since it was last committed. */
     get isDirty(): boolean {
-        return this.committedData && !equal(this.committedData, this.data);
+        return this.committedData && this.committedData !== this.data;
     }
 
     /** Alias for {@link StoreRecord.isDirty} */
@@ -148,24 +148,6 @@ export class StoreRecord {
         return this.validator?.validationState ?? 'Unknown';
     }
 
-    get modifiedData(): PlainObject {
-        if (!this.isModified) return {};
-
-        const ret = this.fields.reduce((acc, {name}) => {
-            if (isNil(this.committedData) || this.data[name] !== this.committedData[name]) {
-                acc[name] = this.data[name];
-            }
-            return acc;
-        }, {});
-
-        return isEmpty(ret)
-            ? ret
-            : {
-                  id: this.id,
-                  ...ret
-              };
-    }
-
     /** Map of field names to list of errors. */
     get errors(): Record<string, string[]> {
         return this.validator?.errors ?? {};
@@ -204,6 +186,26 @@ export class StoreRecord {
     }
 
     /**
+     * Get a map of modified values only.
+     *
+     * If record has no modifications, this method will return null.
+     * If modifications are returned, the returned object will include id,
+     * for convenience.
+     */
+    getModifiedValues(): PlainObject {
+        if (!this.isModified) return null;
+
+        const {data, committedData} = this,
+            ret = pickBy(data, (v, k) => !equal(v, committedData[k]));
+        if (!isEmpty(ret)) {
+            ret.id = this.id;
+            return ret;
+        } else {
+            return null;
+        }
+    }
+
+    /**
      * Construct a StoreRecord from a pre-processed `data` source object.
      *
      * Not typically called by applications directly - `Store` instances create `StoreRecord`s when
@@ -214,15 +216,7 @@ export class StoreRecord {
      * @internal
      */
     constructor(config: StoreRecordConfig) {
-        const {
-            id,
-            store,
-            data,
-            raw = null,
-            committedData = data,
-            parent,
-            isSummary = false
-        } = config;
+        const {id, store, raw, data, committedData, parent, isSummary} = config;
         throwIf(
             isNil(id),
             "Record needs an ID. Use 'Store.idSpec' to specify a unique ID for each record."